+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Version 2.0
 ===========
 - Add API v3.11 support
 - Add informations endpoint support
+- Add capability to inject custom `AuthService` to `ApiInsee` constructor
 - Drop unused exception `ParamsExeption`
 - Replace class and methods violate PEP8 naming convention
     - Errors classes are renamed SubjectError